+++ aquadoggo/src/network/replication/mod.rs
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod behaviour;
+mod handler;
+mod protocol;
+
+pub use behaviour::Behaviour;
+pub use handler::Handler;
+pub use protocol::{Codec, CodecError, Message, Protocol, PROTOCOL_NAME};